set location permission of the user that can access."Chhunhak.CHHOEUNG"

+++ application/controllers/location.php
@@ -137,14 +137,18 @@ class location extends CI_Controller {
 		// delete location by Danet THORNG
 	public function delete_location()
 	{
-
-		$locationID = $this->input->get('loc_id');
-		$this->load->model('Users_model');
-		$data = $this->Users_model->delete_location($locationID);
-		if ($data == 'true') {
-			redirect('location');
+		$user = $this->userlevel();
+		if ($user == 'admin') {
+			$locationID = $this->input->get('loc_id');
+			$this->load->model('Users_model');
+			$data = $this->Users_model->delete_location($locationID);
+			if ($data == 'true') {
+				redirect('location');
+			}else{
+				echo "not delete";
+			}
 		}else{
-			echo "not delete";
+			redirect('errors/error');
 		}
 	}
 }
